Rename queueGCManager.ts to QueueGCManager.ts
* Rename queueGCManager.ts to QueueGCManager.ts
